[updated] 채팅, 중고 상세: 컴포넌트 분리, 리팩토링, 주석
